requirements.txt
flask==3.0.0
flask-login==0.6.3
werkzeug==3.1.3
bcrypt>=4.0.0
pillow>=10.0.0
requests>=2.28.0
pymysql>=1.0.0
python-dotenv==1.0.0
pytz>=2023.3